refactor a tad
