+++ adapter/cd/src/cd/mod.rs
@@ -7,22 +7,15 @@ pub use crate::cd::config::Config;
 mod zpr;
 pub use crate::cd::zpr::Zpr;
 
-use std::{fs, io, sync::Arc};
+mod startmeup;
 
+
+use std::{fs, io, sync::Arc};
 use tracing::{error, info};
 use tracing_subscriber;
+use tokio::signal;
+use tokio::sync::oneshot;
 
-use tokio::{
-    signal,
-    // signal::unix::{signal, SignalKind},
-    //io::{
-    //AsyncReadExt,
-    //AsyncWriteExt,
-    //BufReader,
-    //AsyncBufReadExt
-    //},
-    sync::oneshot,
-};
 
 use tokio_util::{
     sync::CancellationToken,
